# api/contract/http/v1/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Delivery
  description: Отвечает за учет курьеров, диспетчеризацию и доставку
tags:
  - name: Delivery
paths:
  /api/v1/orders:
    post:
      summary: Создать заказ
      description: Позволяет создать заказ с целью тестирования
      operationId: CreateOrder
      responses:
        '201':
          description: Успешный ответ
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/orders/active:
    get:
      summary: Получить все незавершенные заказы
      description: Позволяет получить все незавершенные
      operationId: GetOrders
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/couriers:
    get:
      summary: Получить всех курьеров
      description: Позволяет получить всех курьеров
      operationId: GetCouriers
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Courier'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Location:
      allOf:
        - required:
            - x
            - y
          properties:
            x:
              type: integer
              description: X
            y:
              type: integer
              description: Y
    Order:
      allOf:
        - required:
            - id
            - location
          properties:
            id:
              type: string
              format: uuid
              description: Идентификатор
            location:
              description: Геолокация
              $ref: '#/components/schemas/Location'
    Courier:
      allOf:
        - required:
            - id
            - name
            - location
          properties:
            id:
              type: string
              format: uuid
              description: Идентификатор
            name:
              description: Имя
              type: string
            location:
              description: Геолокация
              $ref: '#/components/schemas/Location'
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Код ошибки
        message:
          type: string
          description: Текст ошибки